middleware auth faille delete

+++ controllers/stuff.js
@@ -21,9 +21,23 @@ exports.modifyThing = (res, req, next) => {
 
 exports.deleteThing = (req, res, next) => {
     // deleteOne va nous permettre de supprimer un élément de notre base de données
-    Thing.deleteOne({ _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-      .catch(error => res.status(400).json({ error }));
+    Thing.findOne({ _id: req.params.id}).then(
+        (thing) => {
+            if (!thing) { //Si on a pas d'objet 
+                return res.status(404).json({
+                    error: new Error('Objet non trouvé')
+                });
+            } // Si l'utilisateur n'est pas le proprietaire de cet objet
+            if (thing.userId !== req.auth.userId) {
+                return res.status(401).json({
+                    error: new Error('requete non authorisé')
+                })
+            }
+            Thing.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
+            .catch(error => res.status(400).json({ error }));
+        }
+    )
 };
 
 exports.getOneThing = (req, res, next) => {
